servlets

+++ servlet/src/main/java/com/geekhub/dao/mapper/GuestBookMapper.java
@@ -11,7 +11,7 @@ public class GuestBookMapper implements RowMapper<GuestBook> {
 
     public GuestBook mapRow(ResultSet rs, int rowNum) throws SQLException {
         GuestBook guestBook = new GuestBook();
-        guestBook.setId(rs.getLong("GuestBook_ID"));
+        guestBook.setId(rs.getInt("GuestBook_ID"));
         guestBook.setName(rs.getString("NameGuest"));
         guestBook.setMessage(rs.getString("Message"));
         guestBook.setRating(rs.getInt("Rating"));
